refactor(neon_framework): let NeonStorage manage the SingleValueStorage
