+++ packages/neon_framework/lib/src/settings/models/storage.dart
@@ -44,43 +44,3 @@ enum StorageKeys implements Storable {
   @override
   final String value;
 }
-
-/// A storage that saves a single value.
-///
-/// [NeonStorage.init] must be called and completed before accessing individual values.
-///
-/// See:
-///   * [NeonStorage] to initialize the storage backend.
-@immutable
-@internal
-final class SingleValueStorage implements KeyValueStorage {
-  /// Creates a new storage for a single value.
-  const SingleValueStorage(this.key);
-
-  @override
-  final StorageKeys key;
-
-  @override
-  bool hasValue() => NeonStorage().database.containsKey(key.value);
-
-  @override
-  Future<bool> remove() => NeonStorage().database.remove(key.value);
-
-  @override
-  String? getString() => NeonStorage().database.getString(key.value);
-
-  @override
-  Future<bool> setString(String value) => NeonStorage().database.setString(key.value, value);
-
-  @override
-  bool? getBool() => NeonStorage().database.getBool(key.value);
-
-  @override
-  Future<bool> setBool(bool value) => NeonStorage().database.setBool(key.value, value);
-
-  @override
-  List<String>? getStringList() => NeonStorage().database.getStringList(key.value);
-
-  @override
-  Future<bool> setStringList(List<String> value) => NeonStorage().database.setStringList(key.value, value);
-}
