+++ packages/neon_framework/lib/src/storage/storage_manager.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 import 'package:meta/meta.dart';
 import 'package:neon_framework/src/settings/models/storage.dart';
 import 'package:neon_framework/src/storage/settings_store.dart';
+import 'package:neon_framework/src/storage/single_value_store.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 /// Neon storage that manages the storage backend.
@@ -59,4 +60,7 @@ class NeonStorage {
 
   /// Initializes a new `SettingsStorage`.
   SettingsStorage settingsStorage(StorageKeys groupKey, [String? suffix]) => AppStorage(groupKey, suffix);
+
+  /// Initializes a new `KeyValueStorage`.
+  KeyValueStorage singleValueStore(StorageKeys key) => SingleValueStorage(key);
 }
