+++ packages/neon_framework/lib/storage.dart
@@ -5,5 +5,5 @@ library;
 
 export 'package:neon_framework/src/storage/request_cache.dart';
 export 'package:neon_framework/src/storage/settings_store.dart' show SettingsStorage;
-export 'package:neon_framework/src/storage/single_value_store.dart';
+export 'package:neon_framework/src/storage/single_value_store.dart' show KeyValueStorage;
 export 'package:neon_framework/src/storage/storage_manager.dart';
